# Reduce allocations while loading classes