@franz1981 2023-09-12
Turning into draft: i need to add a deterministic `free` and correctly handling recursive loading