@franz1981 2023-09-12
@geoand @dmlloyd https://github.com/netty/netty/blob/f1fa227ddf675f055766d04900cce6804fd7f710/common/src/main/java/io/netty/util/internal/CleanerJava9.java#L30

FYI this is what I was referring to, in order to take control of the off-heap ByteBuffer(s) allocated, and keep on reusing it (unless a recursive class loading happen :/ in which case I would leave the default behaviour which use byte[]s and let JNI to malloc/free it)

I would likely use a `MethodHandle::invokeExact` to invoke the Unsafe method to free the buffer, but this can present 2 issues:
- when we allocate/free ourself, we will likely receive some perf penalty while in interpreted mode (on C1 I have to measure it), which pooling would amortize unless we load a single class or the buffer capacity sequence is very unlucky (ie always increasing capacity sizes)
- MH have some inital lookup cost AND I've no idea if the polymorphic signature optimization which avoid allocating the parameters/return value, work as good while in interpreted/C1 compiled: if the pool work as expect this will likely be a very rare event, which will make it more prone to not be optimized at C2 level

Another point to consider related pooling: when the startup is completed I would disable it (including the heap buffer too, probably, unless in the resource(s) loop, local to the method).
To be sure it's the right thing I would like to double check with Hibernate, which has a tangible amount of classes to load (antl included) and could greatly benefit from pooling @Sanne