@franz1981 2023-09-13
yep @dmlloyd my child got sick and today I was forced to park it :"/

Anyway for a "simple" config-quickstart (developement branch) the allocations are reduced from 39 MB to 33 MB, which is a ~15% saving, but it always depends by different factors, really...

I noticed that NettyRecorder (any idea why @geoand ?) load a lot of classes (eg Netty's PlatformDependent) recursively AFTER the startup, which is pretty intensive, and this changes (which pool heap buffers during startup) won't save the day; I have no idea what would happen with hibernate (I have to check) and will make not easy to quantify the saving (of heap) for JVM mode.
In term of "startup" time I was getting a slight regression (probably because the size of the allocated data isn't enough that mem zeroing them is such troublesome), but I was still investigating.

The commit to look at is d3ba38892859cffd2af497920bcc217fa04c4550