@Sanne 2023-09-15
Let's bear in mind that even some code related to bootstrap takes advantage of multiple threads.
I'd say keep this safe in all cases w/o weird context specific shortcuts.