@franz1981 2023-09-15
> I'd say keep this safe in all cases w/o weird context specific shortcuts.

Yep, good advice! I think I'll keep it safe for the main thread during startup (which seems the easier thing to do); do you know if the hibernate classes are loaded during such phase?