@Sanne 2023-09-15
Most Hibernate classes are initialized by a secondary thread, so to not delay the primary thread