@franz1981 2023-09-15
> Most Hibernate classes are initialized by a secondary thread

Ok so it won't benefit from this appraoch :/ that's bad...I need to find a different way and see if that is cool enough