@franz1981 2023-09-18
@geoand @Sanne @dmlloyd Other round...

I've implemented a lock-free intrusive linked-stack to pool buffer spans; it has these characteristics:
- assuming low concurrency (as startup should have), I won't expect problems from compareAndSet retry loops
- but still it perform 2 compare and set for each load: strong volatile-like ops are pretty terrible if performed in tight loops, hence I expect (but we need to measure) that the actual unzip/read jars would dominate the cost between the compareAndSet, making them negligible

I'll both try config-quickstart (which now should work as good as the very first "broken" version) and the hibernate one.

An alternative solution would be to just use a `getAndSet` on acquisition and an weaker "lazySet" (or setRelease in JMM 9) for the release, which should be much faster; this will lessen the pain of having 2 compareAndSet while still getting some weak benefit of reusing buffers.