@franz1981 2023-09-18
So, numbers look great, actually

for config quickstart, it was

> using main thread id confinement: 34.5 MB allocated
without main thread id confinement, still pooling: 33.4 MB allocated
main: 38.9 MB

While here I have 32.7 MB, which is a 6 MB saving of heap allocations