@dmlloyd 2023-09-18
One other thing we looked at, once upon a time, was to write JARs containing uncompressed `.class` files and using `FileChannel.map` (aka `mmap`) to keep the whole JAR mapped in memory, and defining classes based on buffer slices of the main mapped buffer. This would have very little allocation overhead, and should have considerably less copying as well.

It might be worth revisiting at some point.