@geoand 2023-09-19
@dmlloyd @stuartwdouglas would it not be possible to just explode the jars at build time and use directories (a new `DirectoryResource` essentially) in the `target/quarkus-app/lib/main/` instead of the jars?

That would eliminate any unzipping while also allowing for docker to do it's layering AFAIU.

Is there any reason not to follow such an approach?