@franz1981 2023-09-19
@dmlloyd 

> One other thing we looked at, once upon a time, was to write JARs containing uncompressed .class files and using FileChannel.map (aka mmap) to keep the whole JAR mapped in memory, and defining classes based on buffer slices of the main mapped buffer. This would have very little allocation overhead, and should have considerably less copying as well.

This approach is something that if we would be bare-metal would be my first choice, but will concern be about 2 aspects:
- longer time to safe point on normal `ByteBuffer::get` operations: given that the JVM is not in a safepoint while running Java code, if the OS decide to drop/reclaim a page which mmap on jar, it will issue an major page fault on normal java calls which access such memory eg ByteBuffer::get. This will cause stealthy "hiccups" to the application which won't appear in normal logs, see https://ocadaruma.hatenablog.com/entry/2022/12/25/175333 
- RSS and mapped pages: IIRC (and I could be wrong, I've to double check), mapped pages can be accounted as increased RSS which can alert OOM killers, a problem which have affected mmap-based databases like lmdb