@franz1981 2023-09-19
@geoand In my final tests using both the quickstart of panache hibernate and config ones, I had a small (but noticeable) startup improvement, considering the `startup` number collected by quarkus itself (not the time to first request!).

`config quickstart`:
| before  | after | speedup
| ------------- | ------------- |------------- 
|0.335s  | 0.329s | 1.8% 

`panache hibernate quickstart`:

| before  | after | speedup
| ------------- | ------------- 
|0.821s  |0.808s | 1.6%


In short, the speedup exist (no slowdown, which is a good thing although we perform a more complex logic), but the saving in term of allocated memory is more (as explained in https://github.com/quarkusio/quarkus/pull/35875#issuecomment-1723426488)