@franz1981 2023-09-19
Thanks to the suggestion from @Sanne I have a better API in mind...tomorrow going to add a commit with it