@Sanne 2023-09-20
> @franz1981 since you are going to change things, please add the necessary comments in the code as we are sure to forget things in the future and it would be nice to not have to read through all the discussions again :)

right, that's the gist of my "Changes requested" as well ;)